week_three/requirements.txt
# Core dependencies
langchain==0.1.0
langchain-google-genai==0.0.6
streamlit==1.29.0
python-dotenv==1.0.0
google-generativeai==0.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
typing-extensions==4.8.0
pydantic==2.5.0